Move btree unit test to their native, privileged location
